refactor(sketch): sketch 컴포넌트 재사용한 기능별로 분리
